requirements.txt
dlib==19.24.2
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
